// src/app/interfaces/cliente-cadastro.interface.ts
export class IClienteCadastro {
  codCliente: number;
  codigo: number = 0;
  fisJur: string = "F";
  cgcCpf: string = "";
  razao: string = "";
  endereco: string = "";
  numEnd: string = "";
  complemento: string = "";
  bairro: string = "";
  cidade: string = "";
  estado: string = "";
  cep: string = "";
  ddd1: string = "";
  fone1: string = "";
  ddd2: string = "";
  fone2: string = "";
  fax: string = "";
  contato: string = "";
  inscRg: string = "";
  codVen: string = "";
  email: string = "";
  data: Date = new Date();
  status: string = "";
}

export class IClienteCadastroNew {
  codigo: number = 0;
  codRep: number = 0;
  fisJur: string = "F";
  cgcCpf: string = "";
  situacao: string = "A";
  tipo: string = "R";
  razao: string = "";
  endereco: string = "";
  numEnd: string = "";
  complemento: string = "";
  bairro: string = "";
  cidade: string = "";
  estado: string = "";
  cep: string = "";
  ddd1: string = "";
  fone1: string = "";
  ddd2: string = "";
  fone2: string = "";
  fax: string = "";
  contato: string = "";
  datNascCont: string = "";
  dddCont: string = "";
  foneCont: string = "";
  inscRg: string = "";
  consEnt: string = "";
  etiqueta: string = "S";
  consCob: string = "N";
  codAtv: number = 0;
  codNat: number = 0;
  codTra: number = 0;
  codCpg: string = "V01";
  codVen: string = "";
  obsVen: string = "";
  conceito: string = "B";
  bloqueio: string = "";
  limite: number = 0;
  datSci: Date = new Date();
  obsSci: string = "";
  obsCred: string = "";
  obsGer: string = "";
  email: string = "";
  page: string = "";
  datCad: Date = new Date();
  priCpr: Date = new Date();
  ultCpr: Date = new Date();
  datMaiCpr: Date = new Date();
  valMaiCpr: number = 0;
  valMaiAcu: number = 0;
  datMaiAcu: Date = new Date();
  cupomDescto: string = "N";
  cupomValid: Date = new Date();
  cupomFilial: number = 0;
  cupomNota: number = 0;
  cupomModelo: string = "";
  cupomData: Date = new Date();
  regTrib: string = "LP";
  perMvaMt: number = 0;
  perRedMva: number = 0;
  controle: number = 1;
  vendePrazo: string = "S";
  regiao: string = "";
  consEstat: string = "S";
  conheceu: number = 2;
  valorCred: number = 0;
  solicitante: string = "";
  aprovador: string = "";
  prospect: string = "N";
  categoria: string = "P";
  subCatInfo: string = "N";
  subCatBrin: string = "N";
  subCatArte: string = "N";
  subCatPres: string = "N";
  subCatUtil: string = "N";
  subCatAvia: string = "N";
  benefFiscal: string = "BRA";
  suframa: string = "";
  clienteVipi: string = "N";
  data: Date = new Date();
  status: string = "";
}